+++ app/src/main/res/layout/fragment_on_boarding.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.sunul.sunul.presentation.onboarding.OnBoardingVIewModel" />
+            type="com.sunul.sunul.presentation.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
